*: automated-testing -> test, make building test binary optional
package maintainers absolutely don't need the test binary, nor do
users building from the source tarball. we still need to automatically
run the test suite for github actions though.
