+++ test/output.c
@@ -21,7 +21,7 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "automated-testing.h"
+#include "test.h"
 
 #include "headers.h"
 #include "disko.h"
