+++ test/testresult.c
@@ -21,22 +21,17 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "automated-testing.h"
+#include "test.h"
 
-test_index_entry automated_tests[] =
-	{
-#define TEST_FUNC(x) { #x, x },
-#include "test-funcs.h"
-		{0}
-	};
-
-test_index_entry *test_get_case(const char *name)
+// Allow 12 characters
+const char *testresult_str(testresult_t result)
 {
-	int i;
-
-	for (i = 0; automated_tests[i].name; i++)
-		if (!strcmp(automated_tests[i].name, name))
-			return &automated_tests[i];
-
-	return NULL;
+	switch (result) {
+	case SCHISM_TESTRESULT_NOT_RUN: return "NOT RUN";
+	case SCHISM_TESTRESULT_PASS: return "PASS";
+	case SCHISM_TESTRESULT_FAIL: return "FAIL";
+	case SCHISM_TESTRESULT_INCONCLUSIVE: return "INCONCLUSIVE";
+	case SCHISM_TESTRESULT_SKIP: return "SKIP";
+	default: return "#UNKNOWN";
+	}
 }
